tech-debt: close postgres pool in closeTestApp + extract redis-mock (ROK-1058)
The integration test suite leaked the postgres-js connection pool created at
test-app.ts:174 (`max:10` per app instance) because DrizzleModule has no
onModuleDestroy hook — `app.close()` did not cascade. With 49 spec files
× 10 sockets, runs hit Postgres `max_connections` (default 100) and late
files timed out in `beforeAll` with `testApp.db undefined` (ROK-1104).

Stash the raw `appClient` on the `TestApp` interface as `_appClient` and
end it in `closeTestApp` between `app.close()` and `container?.stop()` with
a 5s timeout (postgres-js default 30s would reintroduce ROK-1104 symptoms
at teardown of the last spec).

Extract the redis-mock factory (~80 lines) to `redis-mock.ts` per the
architect's correction #6 to keep `test-app.ts` under the 300-line ESLint
cap as the rest of the ROK-1058 fix lands. Also rewrites the stale
"singleton survives across files" comment block to reflect actual
mechanics (per-file VM context isolation).
